Renaming context directory
