+++ pages/_app.js
@@ -1,5 +1,5 @@
-import { AuthProvider } from '../context/auth';
-import { OverlayProvider } from '../context/overlay';
+import { AuthProvider } from '../contexts/auth';
+import { OverlayProvider } from '../contexts/overlay';
 import { library } from '@fortawesome/fontawesome-svg-core';
 import { fas } from '@fortawesome/free-solid-svg-icons';
 import 'a11y-components/dist/Listbox/Listbox.css';
