+++ pages/login.js
@@ -5,7 +5,7 @@ import { firebaseAPI } from '../lib/firebase';
 import { Button, Input } from 'a11y-components';
 
 import styles from '../styles/Pages.module.scss';
-import { useAuth } from '../context/auth';
+import { useAuth } from '../contexts/auth';
 
 const LoginPage = () => {
   const router = useRouter();
