+++ pages/overlay.js
@@ -1,7 +1,7 @@
 import { useLayoutEffect } from 'react';
 import { Layout, Overlay } from '../components';
-import { useAuth } from '../context/auth';
-import { useOverlay } from '../context/overlay';
+import { useAuth } from '../contexts/auth';
+import { useOverlay } from '../contexts/overlay';
 
 const OverlayPage = ({ fontFamilies }) => {
   const { user, userLoading } = useAuth();
